requirements.txt
gradio==3.50.2
openai==1.3.0
pandas==2.1.1
matplotlib==3.8.0
numpy==1.26.0
python-dotenv==1.0.0 